+++ changes.md
@@ -98,6 +98,7 @@ match(['foo', 'far'])
 - **[new]** .is() - helper-method for comparing two views
 - **[new]** .none() - helper-method for returning an empty view of the document
 - **[new]** .toView() method - drop back to a normal Class instance
+- **[new]** .growLeft() and .growRight() methods
 - **[new]** add punctuation match support via pre/post params
 - **[new]** add ambiguous empty .map() state as 2nd param